refactor: Minimize the set of always-on deps
