+++ protocol/contracts/lpp/Cargo.toml
@@ -9,14 +9,14 @@ license.workspace = true
 crate-type = ["cdylib", "rlib"]
 
 [features]
-net_dev = ["currencies/net_dev"]
-net_test = ["currencies/net_test"]
-net_main = ["currencies/net_main"]
+net_dev = ["contract", "currencies/net_dev"]
+net_test = ["contract", "currencies/net_test"]
+net_main = ["contract", "currencies/net_main"]
 
-astroport = ["currencies/astroport"]
-osmosis = ["currencies/osmosis"]
+astroport = ["contract", "currencies/astroport"]
+osmosis = ["contract", "currencies/osmosis"]
 
-contract = ["stub", "dep:cosmwasm-std", "dep:versioning"]
+contract = ["stub", "sdk/contract", "dep:cosmwasm-std", "dep:versioning"]
 stub = []
 testing = ["currency/testing"]
 
@@ -27,7 +27,7 @@ currency = { workspace = true }
 finance = { workspace = true }
 lpp-platform = { workspace = true }
 platform = { workspace = true }
-sdk = { workspace = true, features = ["contract"] }
+sdk = { workspace = true }
 versioning = { workspace = true, optional = true }
 
 # Required as a dependency by `entry_point` attribute macro
