+++ protocol/contracts/oracle/Cargo.toml
@@ -11,25 +11,31 @@ license.workspace = true
 crate-type = ["cdylib", "rlib"]
 
 [features]
-net_dev = ["currencies/net_dev"]
-net_test = ["currencies/net_test"]
-net_main = ["currencies/net_main"]
+net_dev = ["contract", "currencies/net_dev"]
+net_test = ["contract", "currencies/net_test"]
+net_main = ["contract", "currencies/net_main"]
 
-astroport = ["currencies/astroport"]
-osmosis = ["currencies/osmosis"]
+astroport = ["contract", "currencies/astroport"]
+osmosis = ["contract", "currencies/osmosis"]
 
-contract = ["stub", "dep:cosmwasm-std", "dep:versioning"]
+contract = [
+    "stub",
+    "sdk/contract",
+    "dep:cosmwasm-std",
+    "dep:currencies",
+    "dep:versioning",
+]
 testing = ["stub", "marketprice/testing"]
-stub = [] # TODO strip off any unnecessary dependency
+stub = ["dep:oracle-platform"]
 
 [dependencies]
-currencies = { workspace = true }
+currencies = { workspace = true, optional = true }
 currency = { workspace = true }
 finance = { workspace = true }
 marketprice = { workspace = true }
-oracle-platform = { workspace = true }
+oracle-platform = { workspace = true, optional = true }
 platform = { workspace = true }
-sdk = { workspace = true, features = ["contract"] }
+sdk = { workspace = true }
 swap = { workspace = true }
 tree = { workspace = true, features = ["schema"] }
 versioning = { workspace = true, optional = true, features = ["schema"] }
