+++ protocol/contracts/oracle/src/lib.rs
@@ -1,18 +1,15 @@
 pub use crate::error::ContractError;
 
 pub mod alarms;
+#[cfg(feature = "contract")]
+pub mod contract;
 pub mod error;
+#[cfg(feature = "testing")]
+mod macros;
 pub mod msg;
 pub mod result;
 pub mod state;
-
-#[cfg(any(feature = "contract", test))]
-pub mod contract;
-#[cfg(any(feature = "stub", test))]
+#[cfg(feature = "stub")]
 pub mod stub;
-
-#[cfg(any(test, feature = "testing"))]
-mod macros;
-
 #[cfg(test)]
-pub mod tests;
+mod tests;
