+++ protocol/contracts/oracle/src/tests/mod.rs
@@ -74,7 +74,7 @@ pub(crate) fn dummy_default_instantiate_msg() -> InstantiateMsg {
         StableC1::TICKER.to_string(),
         60,
         Percent::from_percent(50),
-        cosmwasm_std::from_json(format!(
+        sdk::cosmwasm_std::from_json(format!(
             r#"{{
                 "value":[0,"{usdc}"],
                 "children":[
